+++ packages/api/tests/endpoints/test-rules.js
@@ -145,7 +145,8 @@ test.before(async (t) => {
   await ruleModel.createTable();
   t.context.ruleModel = ruleModel;
 
-  const ruleRecord = await ruleModel.create(testRule);
+  const ruleWithTrigger = await ruleModel.createRuleTrigger(testRule);
+  const ruleRecord = await ruleModel.create(ruleWithTrigger);
   await indexer.indexRule(esClient, ruleRecord, esAlias);
 
   const username = randomString();
@@ -428,6 +429,180 @@ test('POST creates a rule', async (t) => {
   );
 });
 
+test.serial('post() creates SNS rule with same trigger information in Dynamo/PostgreSQL', async (t) => {
+  const {
+    pgProvider,
+    pgCollection,
+  } = t.context;
+
+  const topic1 = await awsServices.sns().createTopic({ Name: randomId('topic') }).promise();
+
+  const rule = fakeRuleFactoryV2({
+    state: 'ENABLED',
+    rule: {
+      type: 'sns',
+      value: topic1.TopicArn,
+    },
+    collection: {
+      name: pgCollection.name,
+      version: pgCollection.version,
+    },
+    provider: pgProvider.name,
+  });
+
+  const expressRequest = {
+    body: rule,
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await post(expressRequest, response);
+
+  const dynamoRule = await ruleModel.get({ name: rule.name });
+  const pgRule = await t.context.rulePgModel
+    .get(t.context.testKnex, { name: rule.name });
+
+  t.truthy(dynamoRule.rule.arn);
+  t.like(dynamoRule, {
+    rule: {
+      type: 'sns',
+      value: topic1.TopicArn,
+      arn: dynamoRule.rule.arn,
+    },
+  });
+  t.like(pgRule, {
+    name: rule.name,
+    enabled: true,
+    type: 'sns',
+    arn: dynamoRule.rule.arn,
+    value: topic1.TopicArn,
+  });
+});
+
+test.serial('post() creates the same Kinesis rule with trigger information in Dynamo/PostgreSQL', async (t) => {
+  const {
+    pgProvider,
+    pgCollection,
+  } = t.context;
+
+  const kinesisArn1 = `arn:aws:kinesis:us-east-1:000000000000:${randomId('kinesis')}`;
+  const rule = fakeRuleFactoryV2({
+    state: 'ENABLED',
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn1,
+    },
+    collection: {
+      name: pgCollection.name,
+      version: pgCollection.version,
+    },
+    provider: pgProvider.name,
+  });
+
+  const expressRequest = {
+    body: rule,
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await post(expressRequest, response);
+
+  const dynamoRule = await ruleModel.get({ name: rule.name });
+  const pgRule = await t.context.rulePgModel
+    .get(t.context.testKnex, { name: rule.name });
+
+  t.truthy(dynamoRule.rule.arn);
+  t.truthy(dynamoRule.rule.logEventArn);
+
+  t.like(dynamoRule, {
+    rule: {
+      type: 'kinesis',
+      value: kinesisArn1,
+    },
+  });
+  t.like(pgRule, {
+    name: rule.name,
+    enabled: true,
+    type: 'kinesis',
+    arn: dynamoRule.rule.arn,
+    value: kinesisArn1,
+    log_event_arn: dynamoRule.rule.logEventArn,
+  });
+});
+
+test.serial('post() creates the SQS rule with trigger information in Dynamo/PostgreSQL', async (t) => {
+  const {
+    pgProvider,
+    pgCollection,
+  } = t.context;
+
+  const queue1 = randomId('queue');
+
+  const stubbedRulesModel = new Rule({
+    SqsUtils: {
+      sqsQueueExists: () => Promise.resolve(true),
+    },
+    SqsClient: {
+      getQueueAttributes: () => ({
+        promise: () => Promise.resolve({
+          Attributes: {
+            RedrivePolicy: 'policy',
+            VisibilityTimeout: 10,
+          },
+        }),
+      }),
+    },
+  });
+
+  const rule = fakeRuleFactoryV2({
+    state: 'ENABLED',
+    rule: {
+      type: 'sqs',
+      value: queue1,
+    },
+    collection: {
+      name: pgCollection.name,
+      version: pgCollection.version,
+    },
+    provider: pgProvider.name,
+  });
+
+  const expectedMeta = {
+    visibilityTimeout: 10,
+    retries: 3,
+  };
+
+  const expressRequest = {
+    body: rule,
+    testContext: {
+      ruleModel: stubbedRulesModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await post(expressRequest, response);
+
+  const dynamoRule = await ruleModel.get({ name: rule.name });
+  const pgRule = await t.context.rulePgModel
+    .get(t.context.testKnex, { name: rule.name });
+
+  t.like(dynamoRule, {
+    rule: {
+      type: 'sqs',
+      value: queue1,
+    },
+    meta: expectedMeta,
+  });
+  t.like(pgRule, {
+    name: rule.name,
+    enabled: true,
+    type: 'sqs',
+    value: queue1,
+    meta: expectedMeta,
+  });
+});
+
 test('POST creates a rule in Dynamo and PG with correct timestamps', async (t) => {
   const { newRule } = t.context;
 
@@ -497,7 +672,8 @@ test('POST creates a rule that is enabled by default', async (t) => {
 test('POST returns a 409 response if record already exists', async (t) => {
   const { newRule } = t.context;
 
-  await ruleModel.create(newRule);
+  const ruleWithTrigger = await ruleModel.createRuleTrigger(newRule);
+  await ruleModel.create(ruleWithTrigger);
 
   const response = await request(app)
     .post('/rules')
@@ -640,6 +816,7 @@ test.serial('POST does not write to DynamoDB or RDS if writing to DynamoDB fails
 
   const failingRulesModel = {
     exists: () => false,
+    createRuleTrigger: () => Promise.resolve(newRule),
     create: () => {
       throw new Error('Rule error');
     },
@@ -649,7 +826,7 @@ test.serial('POST does not write to DynamoDB or RDS if writing to DynamoDB fails
     body: newRule,
     testContext: {
       dbClient: testKnex,
-      model: failingRulesModel,
+      ruleModel: failingRulesModel,
     },
   };
 
@@ -676,7 +853,8 @@ test('PUT replaces a rule', async (t) => {
 
   await t.context.testKnex.transaction(async (trx) => {
     await t.context.rulePgModel.create(trx, postgresRule);
-    await ruleModel.create(putTestRule, putTestRule.createdAt);
+    const ruleWithTrigger = await ruleModel.createRuleTrigger(putTestRule);
+    await ruleModel.create(ruleWithTrigger, putTestRule.createdAt);
   });
 
   const updateRule = {