+++ CHANGELOG.md
@@ -26,12 +26,19 @@ aws lambda invoke --function-name $PREFIX-data-migration1 \
 
 ### Changed
 
+- **CUMULUS-2845**
+  - Updated rules model to decouple `createRuleTrigger` from `create`.
+  - Updated rules POST endpoint to call `rulesModel.createRuleTrigger` directly to create rule trigger.
+  - Updated rules PUT endpoints to call `rulesModel.createRuleTrigger` if update fails and reversion needs to occur.
 - **CUMULUS-2846**
   - Updated version of `localstack/localstack` used in local unit testing to `0.11.5`
 
 ### Fixed
 
 - Upgraded lodash to version 4.17.21 to fix vulnerability
+- **CUMULUS-2845**
+  - Fixed bug in POST `/rules` endpoint causing rule records to be created
+  inconsistently in DynamoDB and PostgreSQL
 - **CUMULUS-2846**
   - Fixed logic for `PUT /rules/<name>` endpoint causing rules to be saved
   inconsistently between DynamoDB and PostgreSQL